refactor(plugin-inbox): rename google/people/list → list-groups
The file discovers contact groups, not people directly.
