+++ packages/plugins/plugin-inbox/src/operations/google/people/list-groups.ts
