+++ packages/plugins/plugin-inbox/src/operations/index.ts
@@ -14,7 +14,7 @@ export const InboxOperationHandlerSet = OperationHandlerSet.lazy(
   () => import('./extract-contact'),
   () => import('./google/calendar/list'),
   () => import('./google/calendar/sync'),
-  () => import('./google/people/list'),
+  () => import('./google/people/list-groups'),
   () => import('./google/people/sync'),
   () => import('./google/gmail/send'),
   () => import('./google/gmail/sync'),
